+++ modules/runtime/src/cgpu/build.cgpu.c
@@ -1,3 +1,4 @@
+#include "../pch.h"
 #include "cgpu/cgpu_config.h"
 #ifdef CGPU_USE_VULKAN
     #include "vulkan/vulkan_utils.c"
