+++ modules/runtime/src/cgpu/vulkan/cgpu_vulkan_surfaces.c
@@ -1,3 +1,4 @@
+#include "../../pch.h"
 #include "cgpu/backend/vulkan/cgpu_vulkan_surfaces.h"
 #include "cgpu/backend/vulkan/cgpu_vulkan.h"
 #include "cgpu/extensions/cgpu_vulkan_exts.h"
@@ -25,7 +26,6 @@ void cgpu_free_surface_vulkan(CGPUDeviceId device, CGPUSurfaceId surface)
 }
 
 #if defined(_WIN32) || defined(_WIN64)
-#include "../../platform/windows/winheaders.h"
 
 CGPUSurfaceId cgpu_surface_from_hwnd_vulkan(CGPUDeviceId device, HWND window)
 {
