+++ modules/runtime/src/io/build.io.cpp
@@ -1,3 +1,5 @@
+#include "../pch.hpp"
+
 #include "io.cpp"
 #include "common/io_runner.cpp"
 #include "common/io_resolver.cpp"
