+++ modules/runtime/src/pch.h
@@ -1,0 +1,6 @@
+#pragma once
+#include "platform/debug.h"
+
+#ifdef _WIN32
+#include "platform/windows/winheaders.h"
+#endif
