+++ modules/runtime/src/platform/cpuinfo/impl_x86_windows.c
@@ -26,7 +26,7 @@ static void OverrideOsPreserves(OsPreserves* os_preserves) {
   // No override
 }
 
-#include "../../platform/windows/winheaders.h"
+#include "../../pch.h"
 
 #if defined(CPU_FEATURES_MOCK_CPUID_X86)
 extern bool GetWindowsIsProcessorFeaturePresent(DWORD);
