+++ modules/runtime/src/platform/debug.cpp
@@ -1,9 +1,6 @@
-#include "platform/debug.h"
+#include "../pch.hpp"
 #include "shared_library.cpp"
 #include <stdio.h>
-#if defined(_WIN32)
-#include "../platform/windows/winheaders.h"
-#endif
 
 #if defined(_WIN32)
 void skr_debug_output(const char* msg)
