+++ modules/runtime/src/platform/shared_library.cpp
@@ -1,8 +1,8 @@
+#include "../pch.hpp"
 #include "platform/shared_library.hpp"
 #if defined(SKR_OS_UNIX)
     #include <dlfcn.h>
 #elif defined(SKR_OS_WINDOWS)
-    #include "../platform/windows/winheaders.h"
     #include <platform/filesystem.hpp>
 #endif
