+++ modules/runtime/src/platform/windows/process.cpp
@@ -1,8 +1,6 @@
-#include "winheaders.h"
-
+#include "../../pch.hpp"
 #include "platform/process.h"
 #include "platform/memory.h"
-#include <containers/string.hpp>
 
 typedef struct SProcess
 {
