+++ modules/runtime/src/platform/windows/thread.c
@@ -1,4 +1,4 @@
-#include "winheaders.h"
+#include "../../pch.h"
 #include <process.h> // _beginthreadex
 
 #include "platform/thread.h"
