+++ modules/runtime/src/platform/windows/time.c
@@ -1,4 +1,4 @@
-#include "winheaders.h"
+#include "../pch.h"
 
 #include "platform/time.h"
 #include "platform/debug.h"
